Add experimental option to emit resolver map
Summary:

Test Plan:
