+++ examples/production-app/schema.graphql
@@ -1,6 +1,39 @@
 # Schema generated by Grats (https://grats.capt.dev)
 # Do not manually edit. Regenerate by running `npx grats`.
 """
+Indicates that a position is semantically non null: it is only null if there is a matching error in the `errors` array.
+In all other cases, the position is non-null.
+
+Tools doing code generation may use this information to generate the position as non-null if field errors are handled out of band:
+
+```graphql
+type User {
+    # email is semantically non-null and can be generated as non-null by error-handling clients.
+    email: String @semanticNonNull
+}
+```
+
+The `levels` argument indicates what levels are semantically non null in case of lists:
+
+```graphql
+type User {
+    # friends is semantically non null
+    friends: [User] @semanticNonNull # same as @semanticNonNull(levels: [0])
+
+    # every friends[k] is semantically non null
+    friends: [User] @semanticNonNull(levels: [1])
+
+    # friends as well as every friends[k] is semantically non null
+    friends: [User] @semanticNonNull(levels: [0, 1])
+}
+```
+
+`levels` are zero indexed.
+Passing a negative level or a level greater than the list dimension is an error.
+"""
+directive @semanticNonNull(levels: [Int] = [0]) on FIELD_DEFINITION
+
+"""
 A date and time. Serialized as a Unix timestamp.
 
 **Note**: The `@specifiedBy` directive does not point to a real spec, but is
@@ -51,21 +84,21 @@ input PostContentInput @oneOf {
 }
 
 type CreateLikePayload {
-  post: Post
+  post: Post @semanticNonNull
 }
 
 type CreatePostPayload {
-  post: Post
+  post: Post @semanticNonNull
 }
 
 type CreateUserPayload {
-  user: User
+  user: User @semanticNonNull
 }
 
 """A reaction from a user indicating that they like a post."""
 type Like implements Node {
   """The date and time at which the post was liked."""
-  createdAt: Date
+  createdAt: Date @semanticNonNull
   """
   A globally unique opaque identifier for a node. Can be used to fetch the the
   node with the `node` or `nodes` fields.
@@ -74,9 +107,9 @@ type Like implements Node {
   """
   id: ID!
   """The user who liked the post."""
-  liker: User
+  liker: User @semanticNonNull
   """The post that was liked."""
-  post: Post
+  post: Post @semanticNonNull
 }
 
 type LikeConnection {
@@ -84,40 +117,40 @@ type LikeConnection {
   The total number of likes that post has received.
   **Note:** This is separate from the number of edges currently being read.
   """
-  count: Int
-  edges: [LikeEdge!]
+  count: Int @semanticNonNull
+  edges: [LikeEdge!] @semanticNonNull
   """Convenience field to get the nodes from a connection."""
-  nodes: [Like!]
-  pageInfo: PageInfo
+  nodes: [Like!] @semanticNonNull
+  pageInfo: PageInfo @semanticNonNull
 }
 
 type LikeEdge {
-  cursor: String
-  node: Like
+  cursor: String @semanticNonNull
+  node: Like @semanticNonNull
 }
 
 type Mutation {
   """Like a post. This action is taken as the currently logged in user."""
-  createLike(input: CreateLikeInput!): CreateLikePayload
+  createLike(input: CreateLikeInput!): CreateLikePayload @semanticNonNull
   """Create a new post."""
-  createPost(input: CreatePostInput!): CreatePostPayload
+  createPost(input: CreatePostInput!): CreatePostPayload @semanticNonNull
   """Create a new user."""
-  createUser(input: CreateUserInput!): CreateUserPayload
+  createUser(input: CreateUserInput!): CreateUserPayload @semanticNonNull
 }
 
 type PageInfo {
   endCursor: String
-  hasNextPage: Boolean
-  hasPreviousPage: Boolean
+  hasNextPage: Boolean @semanticNonNull
+  hasPreviousPage: Boolean @semanticNonNull
   startCursor: String
 }
 
 """A blog post."""
 type Post implements Node {
   """The author of the post. This cannot change after the post is created."""
-  author: User
+  author: User @semanticNonNull
   """Content of the post in markdown."""
-  content: String
+  content: String @semanticNonNull
   """
   A globally unique opaque identifier for a node. Can be used to fetch the the
   node with the `node` or `nodes` fields.
@@ -129,38 +162,38 @@ type Post implements Node {
   All the likes this post has received.
   **Note:** You can use this connection to access the number of likes.
   """
-  likes(after: String, before: String, first: Int, last: Int): LikeConnection
+  likes(after: String, before: String, first: Int, last: Int): LikeConnection @semanticNonNull
   """The date and time at which the post was created."""
-  publishedAt: Date
+  publishedAt: Date @semanticNonNull
   """The editor-approved title of the post."""
-  title: String
+  title: String @semanticNonNull
 }
 
 type PostConnection {
-  edges: [PostEdge!]
+  edges: [PostEdge!] @semanticNonNull
   """Convenience field to get the nodes from a connection."""
-  nodes: [Post!]
-  pageInfo: PageInfo
+  nodes: [Post!] @semanticNonNull
+  pageInfo: PageInfo @semanticNonNull
 }
 
 type PostEdge {
-  cursor: String
-  node: Post
+  cursor: String @semanticNonNull
+  node: Post @semanticNonNull
 }
 
 type Query {
   """All likes in the system. Note that there is no guarantee of order."""
-  likes(after: String, before: String, first: Int, last: Int): LikeConnection
+  likes(after: String, before: String, first: Int, last: Int): LikeConnection @semanticNonNull
   """Fetch a single `Node` by its globally unique ID."""
   node(id: ID!): Node
   """Fetch a list of `Node`s by their globally unique IDs."""
-  nodes(ids: [ID!]!): [Node]
+  nodes(ids: [ID!]!): [Node] @semanticNonNull
   """All posts in the system. Note that there is no guarantee of order."""
-  posts(after: String, before: String, first: Int, last: Int): PostConnection
+  posts(after: String, before: String, first: Int, last: Int): PostConnection @semanticNonNull
   """All users in the system. Note that there is no guarantee of order."""
-  users(after: String, before: String, first: Int, last: Int): UserConnection
+  users(after: String, before: String, first: Int, last: Int): UserConnection @semanticNonNull
   """The currently authenticated viewer."""
-  viewer: Viewer
+  viewer: Viewer @semanticNonNull
 }
 
 type Subscription {
@@ -168,7 +201,7 @@ type Subscription {
   Subscribe to likes on a post.
   **Note:** Does not immediately return likes, but rather updates as likes are applied.
   """
-  postLikes(postID: String!): LikeConnection
+  postLikes(postID: String!): LikeConnection @semanticNonNull
 }
 
 type User implements Node {
@@ -180,23 +213,23 @@ type User implements Node {
   """
   id: ID!
   """User's name. **Note:** This field is not guaranteed to be unique."""
-  name: String
+  name: String @semanticNonNull
   """
   All posts written by this user. Note that there is no guarantee of order.
   """
-  posts(after: String, before: String, first: Int, last: Int): PostConnection
+  posts(after: String, before: String, first: Int, last: Int): PostConnection @semanticNonNull
 }
 
 type UserConnection {
-  edges: [UserEdge!]
+  edges: [UserEdge!] @semanticNonNull
   """Convenience field to get the nodes from a connection."""
-  nodes: [User!]
-  pageInfo: PageInfo
+  nodes: [User!] @semanticNonNull
+  pageInfo: PageInfo @semanticNonNull
 }
 
 type UserEdge {
-  cursor: String
-  node: User
+  cursor: String @semanticNonNull
+  node: User @semanticNonNull
 }
 
 """The currently authenticated viewer."""
@@ -207,7 +240,7 @@ type Viewer {
   **Note:** Due to the extreme complexity of this algorithm, it can be slow.
   It is recommended to use `@stream` to avoid blocking the client.
   """
-  feed: [Post!]
+  feed: [Post!] @semanticNonNull
   """The currently authenticated user."""
-  user: User
+  user: User @semanticNonNull
 }
