+++ examples/production-app/schema.json
@@ -1,5 +1,4 @@
 {
-  "$schema": "../../src/resolverSchema.json",
   "types": {
     "CreateLikePayload": {
       "post": {
