+++ examples/production-app/schema.ts
@@ -2,23 +2,21 @@
  * Executable schema generated by Grats (https://grats.capt.dev)
  * Do not manually edit. Regenerate by running `npx grats`.
  */
-import { id as likeIdResolver } from "./graphql/Node";
-import { id as userIdResolver } from "./graphql/Node";
-import { id as postIdResolver } from "./graphql/Node";
-import { nodes as postConnectionNodesResolver } from "./models/PostConnection";
-import { nodes as likeConnectionNodesResolver } from "./models/LikeConnection";
-import { likes as queryLikesResolver } from "./models/LikeConnection";
-import { node as queryNodeResolver } from "./graphql/Node";
-import { nodes as queryNodesResolver } from "./graphql/Node";
-import { posts as queryPostsResolver } from "./models/PostConnection";
-import { nodes as userConnectionNodesResolver } from "./models/UserConnection";
-import { users as queryUsersResolver } from "./models/UserConnection";
+import { defaultFieldResolver, GraphQLSchema, GraphQLObjectType, GraphQLInt, GraphQLList, GraphQLNonNull, GraphQLString, GraphQLScalarType, GraphQLID, GraphQLInterfaceType, GraphQLBoolean, GraphQLInputObjectType } from "graphql";
+import { id as likeIdResolver, id as userIdResolver, id as postIdResolver, node as queryNodeResolver, nodes as queryNodesResolver } from "./graphql/Node";
+import { nodes as postConnectionNodesResolver, posts as queryPostsResolver } from "./models/PostConnection";
+import { nodes as likeConnectionNodesResolver, likes as queryLikesResolver, postLikes as subscriptionPostLikesResolver } from "./models/LikeConnection";
+import { nodes as userConnectionNodesResolver, users as queryUsersResolver } from "./models/UserConnection";
 import { Viewer as queryViewerResolver } from "./models/Viewer";
 import { createLike as mutationCreateLikeResolver } from "./models/Like";
 import { createPost as mutationCreatePostResolver } from "./models/Post";
 import { createUser as mutationCreateUserResolver } from "./models/User";
-import { postLikes as subscriptionPostLikesResolver } from "./models/LikeConnection";
-import { GraphQLSchema, GraphQLObjectType, GraphQLInt, GraphQLList, GraphQLNonNull, GraphQLString, GraphQLScalarType, GraphQLID, GraphQLInterfaceType, GraphQLBoolean, GraphQLInputObjectType } from "graphql";
+async function assertNonNull<T>(value: T | Promise<T>): Promise<T> {
+    const awaited = await value;
+    if (awaited == null)
+        throw new Error("Cannot return null for semantically non-nullable field.");
+    return awaited;
+}
 export function getSchema(): GraphQLSchema {
     const DateType: GraphQLScalarType = new GraphQLScalarType({
         description: "A date and time. Serialized as a Unix timestamp.\n\n**Note**: The `@specifiedBy` directive does not point to a real spec, but is\nincluded here for demonstration purposes.",
@@ -45,12 +43,18 @@ export function getSchema(): GraphQLSchema {
                 author: {
                     description: "The author of the post. This cannot change after the post is created.",
                     name: "author",
-                    type: UserType
+                    type: UserType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 content: {
                     description: "Content of the post in markdown.",
                     name: "content",
-                    type: GraphQLString
+                    type: GraphQLString,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 id: {
                     description: "A globally unique opaque identifier for a node. Can be used to fetch the the\nnode with the `node` or `nodes` fields.\n\nSee: https://graphql.org/learn/global-object-identification/",
@@ -83,18 +87,24 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, _context, info) {
-                        return source.likes(args, info);
+                        return assertNonNull(source.likes(args, info));
                     }
                 },
                 publishedAt: {
                     description: "The date and time at which the post was created.",
                     name: "publishedAt",
-                    type: DateType
+                    type: DateType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 title: {
                     description: "The editor-approved title of the post.",
                     name: "title",
-                    type: GraphQLString
+                    type: GraphQLString,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         },
@@ -108,11 +118,17 @@ export function getSchema(): GraphQLSchema {
             return {
                 cursor: {
                     name: "cursor",
-                    type: GraphQLString
+                    type: GraphQLString,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 node: {
                     name: "node",
-                    type: PostType
+                    type: PostType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -127,11 +143,17 @@ export function getSchema(): GraphQLSchema {
                 },
                 hasNextPage: {
                     name: "hasNextPage",
-                    type: GraphQLBoolean
+                    type: GraphQLBoolean,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 hasPreviousPage: {
                     name: "hasPreviousPage",
-                    type: GraphQLBoolean
+                    type: GraphQLBoolean,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 startCursor: {
                     name: "startCursor",
@@ -146,19 +168,25 @@ export function getSchema(): GraphQLSchema {
             return {
                 edges: {
                     name: "edges",
-                    type: new GraphQLList(new GraphQLNonNull(PostEdgeType))
+                    type: new GraphQLList(new GraphQLNonNull(PostEdgeType)),
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 nodes: {
                     description: "Convenience field to get the nodes from a connection.",
                     name: "nodes",
                     type: new GraphQLList(new GraphQLNonNull(PostType)),
                     resolve(source) {
-                        return postConnectionNodesResolver(source);
+                        return assertNonNull(postConnectionNodesResolver(source));
                     }
                 },
                 pageInfo: {
                     name: "pageInfo",
-                    type: PageInfoType
+                    type: PageInfoType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -178,7 +206,10 @@ export function getSchema(): GraphQLSchema {
                 name: {
                     description: "User's name. **Note:** This field is not guaranteed to be unique.",
                     name: "name",
-                    type: GraphQLString
+                    type: GraphQLString,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 posts: {
                     description: "All posts written by this user. Note that there is no guarantee of order.",
@@ -203,7 +234,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, _context, info) {
-                        return source.posts(args, info);
+                        return assertNonNull(source.posts(args, info));
                     }
                 }
             };
@@ -220,7 +251,10 @@ export function getSchema(): GraphQLSchema {
                 createdAt: {
                     description: "The date and time at which the post was liked.",
                     name: "createdAt",
-                    type: DateType
+                    type: DateType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 id: {
                     description: "A globally unique opaque identifier for a node. Can be used to fetch the the\nnode with the `node` or `nodes` fields.\n\nSee: https://graphql.org/learn/global-object-identification/",
@@ -233,12 +267,18 @@ export function getSchema(): GraphQLSchema {
                 liker: {
                     description: "The user who liked the post.",
                     name: "liker",
-                    type: UserType
+                    type: UserType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 post: {
                     description: "The post that was liked.",
                     name: "post",
-                    type: PostType
+                    type: PostType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         },
@@ -252,11 +292,17 @@ export function getSchema(): GraphQLSchema {
             return {
                 cursor: {
                     name: "cursor",
-                    type: GraphQLString
+                    type: GraphQLString,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 node: {
                     name: "node",
-                    type: LikeType
+                    type: LikeType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -268,23 +314,32 @@ export function getSchema(): GraphQLSchema {
                 count: {
                     description: "The total number of likes that post has received.\n**Note:** This is separate from the number of edges currently being read.",
                     name: "count",
-                    type: GraphQLInt
+                    type: GraphQLInt,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 edges: {
                     name: "edges",
-                    type: new GraphQLList(new GraphQLNonNull(LikeEdgeType))
+                    type: new GraphQLList(new GraphQLNonNull(LikeEdgeType)),
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 nodes: {
                     description: "Convenience field to get the nodes from a connection.",
                     name: "nodes",
                     type: new GraphQLList(new GraphQLNonNull(LikeType)),
                     resolve(source) {
-                        return likeConnectionNodesResolver(source);
+                        return assertNonNull(likeConnectionNodesResolver(source));
                     }
                 },
                 pageInfo: {
                     name: "pageInfo",
-                    type: PageInfoType
+                    type: PageInfoType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -295,11 +350,17 @@ export function getSchema(): GraphQLSchema {
             return {
                 cursor: {
                     name: "cursor",
-                    type: GraphQLString
+                    type: GraphQLString,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 node: {
                     name: "node",
-                    type: UserType
+                    type: UserType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -310,19 +371,25 @@ export function getSchema(): GraphQLSchema {
             return {
                 edges: {
                     name: "edges",
-                    type: new GraphQLList(new GraphQLNonNull(UserEdgeType))
+                    type: new GraphQLList(new GraphQLNonNull(UserEdgeType)),
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 },
                 nodes: {
                     description: "Convenience field to get the nodes from a connection.",
                     name: "nodes",
                     type: new GraphQLList(new GraphQLNonNull(UserType)),
                     resolve(source) {
-                        return userConnectionNodesResolver(source);
+                        return assertNonNull(userConnectionNodesResolver(source));
                     }
                 },
                 pageInfo: {
                     name: "pageInfo",
-                    type: PageInfoType
+                    type: PageInfoType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -337,7 +404,7 @@ export function getSchema(): GraphQLSchema {
                     name: "feed",
                     type: new GraphQLList(new GraphQLNonNull(PostType)),
                     resolve(source, _args, context) {
-                        return source.feed(context);
+                        return assertNonNull(source.feed(context));
                     }
                 },
                 user: {
@@ -345,7 +412,7 @@ export function getSchema(): GraphQLSchema {
                     name: "user",
                     type: UserType,
                     resolve(source, _args, context) {
-                        return source.user(context);
+                        return assertNonNull(source.user(context));
                     }
                 }
             };
@@ -378,7 +445,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, context, info) {
-                        return queryLikesResolver(source, args, context, info);
+                        return assertNonNull(queryLikesResolver(source, args, context, info));
                     }
                 },
                 node: {
@@ -406,7 +473,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, context) {
-                        return queryNodesResolver(source, args, context);
+                        return assertNonNull(queryNodesResolver(source, args, context));
                     }
                 },
                 posts: {
@@ -432,7 +499,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, context, info) {
-                        return queryPostsResolver(source, args, context, info);
+                        return assertNonNull(queryPostsResolver(source, args, context, info));
                     }
                 },
                 users: {
@@ -458,7 +525,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, context, info) {
-                        return queryUsersResolver(source, args, context, info);
+                        return assertNonNull(queryUsersResolver(source, args, context, info));
                     }
                 },
                 viewer: {
@@ -466,7 +533,7 @@ export function getSchema(): GraphQLSchema {
                     name: "viewer",
                     type: ViewerType,
                     resolve(source) {
-                        return queryViewerResolver.viewer(source);
+                        return assertNonNull(queryViewerResolver.viewer(source));
                     }
                 }
             };
@@ -478,7 +545,10 @@ export function getSchema(): GraphQLSchema {
             return {
                 post: {
                     name: "post",
-                    type: PostType
+                    type: PostType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -500,7 +570,10 @@ export function getSchema(): GraphQLSchema {
             return {
                 post: {
                     name: "post",
-                    type: PostType
+                    type: PostType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -584,7 +657,10 @@ export function getSchema(): GraphQLSchema {
             return {
                 user: {
                     name: "user",
-                    type: UserType
+                    type: UserType,
+                    resolve(source, args, context, info) {
+                        return assertNonNull(defaultFieldResolver(source, args, context, info));
+                    }
                 }
             };
         }
@@ -615,7 +691,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, context) {
-                        return mutationCreateLikeResolver(source, args.input, context);
+                        return assertNonNull(mutationCreateLikeResolver(source, args.input, context));
                     }
                 },
                 createPost: {
@@ -629,7 +705,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, context) {
-                        return mutationCreatePostResolver(source, args.input, context);
+                        return assertNonNull(mutationCreatePostResolver(source, args.input, context));
                     }
                 },
                 createUser: {
@@ -643,7 +719,7 @@ export function getSchema(): GraphQLSchema {
                         }
                     },
                     resolve(source, args, context) {
-                        return mutationCreateUserResolver(source, args.input, context);
+                        return assertNonNull(mutationCreateUserResolver(source, args.input, context));
                     }
                 }
             };
@@ -667,7 +743,7 @@ export function getSchema(): GraphQLSchema {
                         return subscriptionPostLikesResolver(source, args.postID, context, info);
                     },
                     resolve(payload) {
-                        return payload;
+                        return assertNonNull(payload);
                     }
                 }
             };
