+++ src/codegen/TSAstBuilder.ts
@@ -1,5 +1,8 @@
 import * as ts from "typescript";
-import { isNonNull } from "./helpers";
+import { isNonNull } from "../utils/helpers";
+import * as path from "path";
+import { resolveRelativePath } from "../gratsRoot";
+
 const F = ts.factory;
 
 /**
@@ -7,11 +10,14 @@ const F = ts.factory;
  */
 export default class TSAstBuilder {
   _imports: ts.Statement[] = [];
+  imports: Map<string, { name: string; as?: string }[]> = new Map();
   _helpers: ts.Statement[] = [];
   _statements: ts.Statement[] = [];
-  addImport(importStatement: ts.Statement) {
-    this._imports.push(importStatement);
-  }
+
+  constructor(
+    private _destination: string,
+    private importModuleSpecifierEnding: string,
+  ) {}
   addHelper(statement: ts.Statement) {
     this._helpers.push(statement);
   }
@@ -108,42 +114,50 @@ export default class TSAstBuilder {
   }
 
   import(from: string, names: { name: string; as?: string }[]) {
-    const namedImports = names.map((name) => {
-      if (name.as) {
-        return F.createImportSpecifier(
-          false,
-          F.createIdentifier(name.name),
-          F.createIdentifier(name.as),
-        );
-      } else {
-        return F.createImportSpecifier(
-          false,
-          undefined,
-          F.createIdentifier(name.name),
-        );
+    let moduleImports = this.imports.get(from);
+    if (moduleImports == null) {
+      moduleImports = [];
+      this.imports.set(from, moduleImports);
+    }
+    for (const { name, as } of names) {
+      let seen = false;
+      for (const imp of moduleImports) {
+        if (imp.name === name && imp.as === as) {
+          seen = true;
+        }
       }
-    });
-    this.addImport(
-      F.createImportDeclaration(
-        undefined,
-        F.createImportClause(
-          false,
-          undefined,
-          F.createNamedImports(namedImports),
-        ),
-        F.createStringLiteral(from),
-      ),
-    );
+      if (!seen) {
+        moduleImports.push({ name, as });
+      }
+    }
   }
 
   importDefault(from: string, as: string) {
-    this.addImport(
+    this._imports.push(
       F.createImportDeclaration(
         undefined,
         F.createImportClause(false, F.createIdentifier(as), undefined),
         F.createStringLiteral(from),
       ),
     );
+  }
+
+  importUserConstruct(
+    tsModulePath: string,
+    exportName: string | null,
+    localName: string,
+  ): void {
+    const abs = resolveRelativePath(tsModulePath);
+    const relative = replaceExt(
+      path.relative(path.dirname(this._destination), abs),
+      this.importModuleSpecifierEnding,
+    );
+    const modulePath = `./${normalizeRelativePathToPosix(relative)}`;
+    if (exportName == null) {
+      this.importDefault(modulePath, localName);
+    } else {
+      this.import(modulePath, [{ name: exportName, as: localName }]);
+    }
   }
 
   print(): string {
@@ -156,6 +170,35 @@ export default class TSAstBuilder {
       ts.ScriptKind.TS,
     );
 
+    for (const [from, names] of this.imports) {
+      const namedImports = names.map((name) => {
+        if (name.as) {
+          return F.createImportSpecifier(
+            false,
+            F.createIdentifier(name.name),
+            F.createIdentifier(name.as),
+          );
+        } else {
+          return F.createImportSpecifier(
+            false,
+            undefined,
+            F.createIdentifier(name.name),
+          );
+        }
+      });
+      this._imports.push(
+        F.createImportDeclaration(
+          undefined,
+          F.createImportClause(
+            false,
+            undefined,
+            F.createNamedImports(namedImports),
+          ),
+          F.createStringLiteral(from),
+        ),
+      );
+    }
+
     return printer.printList(
       ts.ListFormat.MultiLine,
       F.createNodeArray([
@@ -166,4 +209,13 @@ export default class TSAstBuilder {
       sourceFile,
     );
   }
+}
+function replaceExt(filePath: string, newSuffix: string): string {
+  const ext = path.extname(filePath);
+  return filePath.slice(0, -ext.length) + newSuffix;
+}
+
+// https://github.com/sindresorhus/slash/blob/98b618f5a3bfcb5dd374b204868818845b87bb2f/index.js#L8C9-L8C33
+function normalizeRelativePathToPosix(unknownPath: string): string {
+  return unknownPath.replace(/\\/g, "/");
 }
