+++ src/tests/fixtures/extend_interface/addStringFieldToInterface.ts.expected
@@ -48,8 +48,7 @@ type User implements IPerson {
   hello: String
 }
 -- TypeScript --
-import { greeting as adminGreetingResolver } from "./addStringFieldToInterface";
-import { greeting as userGreetingResolver } from "./addStringFieldToInterface";
+import { greeting as adminGreetingResolver, greeting as userGreetingResolver } from "./addStringFieldToInterface";
 import { GraphQLSchema, GraphQLInterfaceType, GraphQLString, GraphQLObjectType } from "graphql";
 export function getSchema(): GraphQLSchema {
     const IPersonType: GraphQLInterfaceType = new GraphQLInterfaceType({
