+++ src/tests/fixtures/todo/userExample.ts.expected
@@ -36,8 +36,7 @@ type User {
   lastName: String
 }
 -- TypeScript --
-import { fullName as userFullNameResolver } from "./userExample";
-import { me as someTypeMeResolver } from "./userExample";
+import { fullName as userFullNameResolver, me as someTypeMeResolver } from "./userExample";
 import { GraphQLSchema, GraphQLObjectType, GraphQLString } from "graphql";
 export function getSchema(): GraphQLSchema {
     const UserType: GraphQLObjectType = new GraphQLObjectType({
