+++ src/tests/integrationFixtures/complexPlurals/schema.ts
@@ -1,6 +1,4 @@
-import { arrayOfArrayOfPromises as queryArrayOfArrayOfPromisesResolver } from "./index";
-import { arrayOfPromises as queryArrayOfPromisesResolver } from "./index";
-import { asyncIterableOfArrayOfPromises as queryAsyncIterableOfArrayOfPromisesResolver } from "./index";
+import { arrayOfArrayOfPromises as queryArrayOfArrayOfPromisesResolver, arrayOfPromises as queryArrayOfPromisesResolver, asyncIterableOfArrayOfPromises as queryAsyncIterableOfArrayOfPromisesResolver } from "./index";
 import { GraphQLSchema, GraphQLObjectType, GraphQLList, GraphQLNonNull, GraphQLString } from "graphql";
 export function getSchema(): GraphQLSchema {
     const QueryType: GraphQLObjectType = new GraphQLObjectType({
