+++ src/tests/integrationFixtures/deprecated/schema.ts
@@ -1,5 +1,4 @@
-import { goodBye as queryGoodByeResolver } from "./index";
-import { hello as queryHelloResolver } from "./index";
+import { goodBye as queryGoodByeResolver, hello as queryHelloResolver } from "./index";
 import { GraphQLSchema, GraphQLObjectType, GraphQLString } from "graphql";
 export function getSchema(): GraphQLSchema {
     const QueryType: GraphQLObjectType = new GraphQLObjectType({
