+++ src/tests/integrationFixtures/nonNullableListIncludesNull/schema.ts
@@ -1,5 +1,4 @@
-import { someList as querySomeListResolver } from "./index";
-import { someListOfLists as querySomeListOfListsResolver } from "./index";
+import { someList as querySomeListResolver, someListOfLists as querySomeListOfListsResolver } from "./index";
 import { GraphQLSchema, GraphQLObjectType, GraphQLList, GraphQLNonNull, GraphQLString } from "graphql";
 export function getSchema(): GraphQLSchema {
     const QueryType: GraphQLObjectType = new GraphQLObjectType({
